refactor(cli): extract command handler boundaries
