+++ crates/git-smee-cli/src/commands/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod init;
+pub(crate) mod install;
+pub(crate) mod run;
